extracts fake https client to test helpers
